requirements.txt
amqp==2.2.2
appnope==0.1.0
attrdict==2.0.0
behave==1.2.6
behave-django==1.1.0
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cytoolz==0.9.0
decorator==4.2.1
defusedxml==0.5.0
Django==2.0.2
django-object-actions==1.0.0
django-webpack-loader==0.5.0
djangorestframework==3.7.7
eth-abi==1.0.0b1
eth-account==0.1.0a2
eth-hash==0.1.0a4
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.0a2
eth-utils==1.0.0b2
factory-boy==2.10.0
Faker==0.8.12
hexbytes==0.1.0b1
honcho==1.0.1
idna==2.6
ipdb==0.11
ipython==6.2.1
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.11.1
Jinja2==2.10
kombu==4.1.0
lru-dict==1.1.6
MarkupSafe==1.0
mysqlclient==1.3.12
oauthlib==2.0.6
parse==1.8.2
parse-type==0.4.2
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycryptodome==3.4.11
pyflakes==1.6.0
Pygments==2.2.0
PyJWT==1.6.1
pysha3==1.0.2
python-dateutil==2.7.0
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
rlp==0.6.0
selenium==3.11.0
simplegeneric==0.8.1
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
text-unidecode==1.2
toolz==0.9.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
web3==4.1.0
pyjwkest==1.4.0
oslash==0.5.1
